Moved the output's tests to their own file
